Change go mod (#9)
* common/types: add Address.Big
Many of the other types have a function to convert the type to a big.Int,
but Address was missing this function.

It is useful to be able to turn an Address into a big.Int when doing
EVM-like computations natively in Go. Sometimes a Solidity address
type is casted to a uint256 and having a Big method on the Address
type makes this easy.

* all: Modify the package name

* cmd/clef: list accounts at startup
Reports accounts known to Clef during startup, after master seed is provided by the user.

---------

Co-authored-by: lukads12345 <lukads12345>